split up all of main covector run code
