+++ packages/covector/bin/covector.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 const { main } = require("effection");
-const { cli, covector } = require("covector");
+const { cli } = require("covector");
 
 main(function* start() {
-  yield cli(process.argv.slice(2), covector);
+  yield cli(process.argv.slice(2));
 });
