+++ packages/covector/src/run.ts
@@ -1,50 +1,25 @@
-import {
-  attemptCommands,
-  confirmCommandsToRun,
-  raceTime,
-} from "@covector/command";
-import {
-  configFile,
-  readPreFile,
-  changeFiles,
-  loadChangeFiles,
-  changeFilesRemove,
-  writePreFile,
-} from "@covector/files";
-import {
-  assemble,
-  mergeIntoConfig,
-  mergeChangesToConfig,
-} from "@covector/assemble";
-import {
-  fillChangelogs,
-  pullLastChangelog,
-  pipeChangelogToCommands,
-} from "@covector/changelog";
-import {
-  apply,
-  changesConsideringParents,
-  validateApply,
-} from "@covector/apply";
-
-import type {
-  CommandsRan,
-  CovectorStatus,
-  CovectorVersion,
-  CovectorPublish,
-  Covector,
-  PkgVersion,
-  PkgPublish,
-} from "@covector/types";
+import { init } from "./init";
+import { add } from "./add";
+import { status } from "./status";
+import { config } from "./config";
+import { version } from "./version";
+import { preview } from "./preview";
+import { publish } from "./publish";
+import { arbitrary } from "./arbitrary";
 
 export function* covector({
+  // shared
   command,
-  dryRun = false,
   cwd = process.cwd(),
+  // usage inputs
+  dryRun = false,
   filterPackages = [],
   modifyConfig = async (c) => c,
   previewVersion = "",
   branchTag = "",
+  // setup inputs
+  changeFolder = ".changes",
+  yes = false,
 }: {
   command: string;
   dryRun?: boolean;
@@ -53,468 +28,57 @@ export function* covector({
   modifyConfig?: (c: any) => Promise<any>;
   previewVersion?: string;
   branchTag?: string;
-}): Generator<any, Covector, any> {
-  const config = yield modifyConfig(yield configFile({ cwd }));
-  const pre = yield readPreFile({ cwd, changeFolder: config.changeFolder });
-  const prereleaseIdentifier = !pre ? null : pre.tag;
-
-  const changesPaths = yield changeFiles({
-    cwd,
-    changeFolder: config.changeFolder,
-  });
-  const changeFilesLoaded = yield loadChangeFiles({
-    cwd,
-    paths: changesPaths,
-  });
-  const assembledChanges = yield assemble({
-    cwd,
-    files: changeFilesLoaded,
-    config,
-    preMode: { on: !!pre, prevFiles: !pre ? [] : pre.changes },
-  });
-
-  if (command === "status" || !command) {
-    if (changeFilesLoaded.length === 0) {
-      console.info("There are no changes.");
-
-      const { commands: publishCommands }: { commands: PkgPublish[] } =
-        yield mergeIntoConfig({
-          assembledChanges,
-          config,
-          command: "publish",
-          cwd,
-          dryRun,
-          filterPackages,
-          tag: branchTag,
-        });
-
-      if (publishCommands.length === 0) {
-        console.log(`No commands configured to run on publish.`);
-        return {
-          response: `No commands configured to run on publish.`,
-          pkgReadyToPublish: [],
-        };
-      }
-
-      const commandsToRun: PkgPublish[] = yield confirmCommandsToRun({
-        cwd,
-        commands: publishCommands,
-        command: "publish",
-      });
-
-      if (commandsToRun.length > 0) {
-        console.log(
-          `There ${
-            commandsToRun.length === 1
-              ? `is 1 package`
-              : `is ${commandsToRun.length} packages`
-          } ready to publish which includes${commandsToRun.map(
-            (pkg) => ` ${pkg.pkg}@${pkg.pkgFile?.version}`
-          )}`
-        );
-      }
-
-      return <CovectorStatus>{
-        pkgReadyToPublish: commandsToRun,
-        response: "No changes.",
-      };
-    } else if (!!pre && assembledChanges?.changes?.length === 0) {
-      console.info("There are no changes.");
-      console.log(
-        "We have previously released the changes in these files:",
-        changesPaths
-      );
-      return { pkgReadyToPublish: [], response: "No changes." };
-    } else {
-      // write out all of the changes
-      // TODO make it pretty
-      console.log("changes:");
-      Object.keys(assembledChanges.releases).forEach((release) => {
-        console.log(`${release} => ${assembledChanges.releases[release].type}`);
-        console.dir(assembledChanges.releases[release].changes);
-      });
-
-      const changes = changesConsideringParents({
-        assembledChanges,
-        config,
-        prereleaseIdentifier,
-      });
-
-      const { commands } = yield mergeChangesToConfig({
-        assembledChanges: changes,
-        config,
-        command,
-        dryRun,
-        filterPackages,
-        cwd,
-      });
-
-      const applied = yield validateApply({
-        commands,
-        config,
-        cwd,
-        prereleaseIdentifier,
-      });
-
-      return <CovectorStatus>{
-        response: `There are ${
-          Object.keys(assembledChanges.releases).length
-        } changes which include${Object.keys(assembledChanges.releases).map(
-          (release) =>
-            ` ${release} with ${assembledChanges.releases[release].type}`
-        )}`,
-      };
-    }
+  changeFolder?: string;
+  yes?: boolean;
+}): Generator<any, any, any> {
+  if (command === "init") {
+    return yield init({ cwd, changeFolder, yes });
+  } else if (command === "add") {
+    return yield add({ cwd, changeFolder, yes });
   } else if (command === "config") {
-    delete config.vfile;
-    console.dir(config);
-    return { response: "config returned" };
+    return yield config({ cwd, modifyConfig });
+  } else if (command === "status") {
+    return yield status({
+      command,
+      dryRun,
+      cwd,
+      filterPackages,
+      modifyConfig,
+      branchTag,
+    });
   } else if (command === "version") {
-    yield raceTime({ t: config.timeout });
-    const changes = changesConsideringParents({
-      assembledChanges,
-      config,
-      prereleaseIdentifier,
-    });
-
-    const {
-      commands,
-      pipeTemplate,
-    }: {
-      commands: PkgVersion[];
-      pipeTemplate: any;
-    } = yield mergeChangesToConfig({
-      assembledChanges: changes,
-      config,
+    return yield version({
       command,
       dryRun,
+      cwd,
       filterPackages,
-      cwd,
+      modifyConfig,
     });
-    if (dryRun) {
-      console.log("==== commands ready to run ===");
-      console.log(commands);
-    }
-
-    let pkgCommandsRan: CommandsRan = Object.keys(config.packages).reduce(
-      (
-        pkgs: {
-          [k: string]: {
-            precommand: string | false;
-            command: string | false;
-            postcommand: string | false;
-            applied: string | false;
-          };
-        },
-        pkg: string
-      ) => {
-        pkgs[pkg] = {
-          precommand: false,
-          command: false,
-          postcommand: false,
-          applied: false,
-        };
-        return pkgs;
-      },
-      {}
-    );
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands,
-      commandPrefix: "pre",
-      command,
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    const applied = yield apply({
-      //@ts-ignore
-      commands,
-      config,
-      cwd,
-      bump: !dryRun,
-      prereleaseIdentifier,
-    });
-
-    pkgCommandsRan = applied.reduce(
-      (
-        pkgs: {
-          [k: string]: {
-            precommand: boolean;
-            command: boolean;
-            postcommand: boolean;
-            applied: object;
-          };
-        },
-        result: { name: string }
-      ) => {
-        pkgs[result.name].applied = result;
-        return pkgs;
-      },
-      pkgCommandsRan
-    );
-
-    pkgCommandsRan = yield fillChangelogs({
-      applied,
-      //@ts-ignore
-      assembledChanges: changes,
-      config,
-      cwd,
-      pkgCommandsRan,
-      create: !dryRun,
-    });
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands,
-      commandPrefix: "post",
-      command,
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    if (command === "version" && !dryRun) {
-      if (pre) {
-        pre.changes = changesPaths;
-        yield writePreFile({ preFile: pre, cwd });
-      } else {
-        yield changeFilesRemove({ cwd, paths: changesPaths });
-      }
-    }
-
-    if (dryRun) {
-      console.log("==== result ===");
-      console.log(pkgCommandsRan);
-    }
-
-    return <CovectorVersion>{ commandsRan: pkgCommandsRan, pipeTemplate };
   } else if (command === "preview") {
-    yield raceTime({ t: config.timeout });
-
-    const versionChanges = changesConsideringParents({
-      assembledChanges,
-      config,
-      prereleaseIdentifier,
-    });
-
-    const { commands: versionCommands } = yield mergeChangesToConfig({
-      assembledChanges: versionChanges,
-      config,
-      command: "version",
+    return yield preview({
+      command,
       dryRun,
+      cwd,
       filterPackages,
-      cwd,
-    });
-
-    let pkgCommandsRan: CommandsRan = Object.keys(config.packages).reduce(
-      (
-        pkgs: {
-          [k: string]: {
-            precommand: string | false;
-            command: string | false;
-            postcommand: string | false;
-            applied: string | false;
-          };
-        },
-        pkg: string
-      ) => {
-        pkgs[pkg] = {
-          precommand: false,
-          command: false,
-          postcommand: false,
-          applied: false,
-        };
-        return pkgs;
-      },
-      {}
-    );
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: versionCommands,
-      commandPrefix: "pre",
-      command: "version",
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    const applied = yield apply({
-      commands: versionCommands,
-      config,
-      cwd,
-      bump: !dryRun,
+      modifyConfig,
       previewVersion,
-      prereleaseIdentifier,
+      branchTag,
     });
-
-    pkgCommandsRan = applied.reduce(
-      (
-        pkgs: {
-          [k: string]: {
-            precommand: boolean;
-            command: boolean;
-            postcommand: boolean;
-            applied: object;
-          };
-        },
-        result: { name: string }
-      ) => {
-        pkgs[result.name].applied = result;
-        return pkgs;
-      },
-      pkgCommandsRan
-    );
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: versionCommands,
-      commandPrefix: "post",
-      command: "version",
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    const { commands: publishCommands }: { commands: PkgPublish[] } =
-      yield mergeIntoConfig({
-        assembledChanges,
-        config,
-        command: "publish",
-        cwd,
-        dryRun,
-        filterPackages,
-        tag: branchTag,
-      });
-
-    if (publishCommands.length === 0) {
-      console.log(`No commands configured to run on publish.`);
-      return {
-        response: `No commands configured to run on publish.`,
-      };
-    }
-
-    const commandsToRun: PkgPublish[] = yield confirmCommandsToRun({
-      cwd,
-      commands: publishCommands,
-      command: "publish",
-    });
-
-    pkgCommandsRan = publishCommands.reduce(
-      (pkgs: any, pkg: { pkg: string }): object => {
-        pkgs[pkg.pkg] = {
-          precommand: false,
-          command: false,
-          postcommand: false,
-          pkg,
-        };
-        return pkgs;
-      },
-      {}
-    );
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: commandsToRun,
-      commandPrefix: "pre",
-      command: "publish",
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: commandsToRun,
-      command: "publish",
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    return { commandsRan: pkgCommandsRan, pipeTemplate: publishCommands };
-  } else {
-    yield raceTime({ t: config.timeout });
-    const changelogs = yield pullLastChangelog({
-      config,
-      cwd,
-    });
-
-    const {
-      commands,
-      pipeTemplate,
-    }: { commands: PkgPublish[]; pipeTemplate: any } = yield mergeIntoConfig({
-      assembledChanges,
-      config,
+  } else if (command === "publish") {
+    return yield publish({
       command,
-      cwd,
       dryRun,
+      cwd,
       filterPackages,
-      changelogs,
+      modifyConfig,
     });
-
-    if (dryRun) {
-      console.log("==== commands ready to run ===");
-      console.log(commands);
-    }
-
-    if (commands.length === 0) {
-      console.log(`No commands configured to run on [${command}].`);
-      return {
-        response: `No commands configured to run on [${command}].`,
-      };
-    }
-
-    const commandsToRun: PkgPublish[] = yield confirmCommandsToRun({
-      cwd,
-      commands,
+  } else {
+    return yield arbitrary({
       command,
-    });
-
-    let pkgCommandsRan: CommandsRan = commands.reduce(
-      (pkgs: any, pkg: { pkg: string }): object => {
-        pkgs[pkg.pkg] = {
-          precommand: false,
-          command: false,
-          postcommand: false,
-          pkg,
-        };
-        return pkgs;
-      },
-      {}
-    );
-
-    pkgCommandsRan = yield pipeChangelogToCommands({
-      changelogs,
-      pkgCommandsRan,
-    });
-
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: commandsToRun,
-      commandPrefix: "pre",
-      command,
-      pkgCommandsRan,
       dryRun,
-    });
-    pkgCommandsRan = yield attemptCommands({
       cwd,
-      commands: commandsToRun,
-      command,
-      pkgCommandsRan,
-      dryRun,
+      filterPackages,
+      modifyConfig,
     });
-    pkgCommandsRan = yield attemptCommands({
-      cwd,
-      commands: commandsToRun,
-      commandPrefix: "post",
-      command,
-      pkgCommandsRan,
-      dryRun,
-    });
-
-    if (dryRun) {
-      console.log("==== result ===");
-      console.log(pkgCommandsRan);
-    }
-
-    return <CovectorPublish>{ commandsRan: pkgCommandsRan, pipeTemplate };
   }
 }
